+++ cli.js
@@ -3,36 +3,21 @@
 var posthtml = require('posthtml');
 var fs = require('fs');
 var argv = require('yargs')
-	.usage('Usage: $0 [--config|-c config.json] [--output|-o output.html] [--input|-i input.html]')
-	.example('posthtml -o output.html input.html', 'Default example')
-	.pkgConf('posthtml')
-	.config('c', {
-		config: true
-	})
-	.alias('c', 'config')
+	.usage('Usage: $0 --output|-o output.html --input|-i input.html [--config|-c config.json]')
+	.example('posthtml -o output.html -i input.html', 'Default example')
+	.demand(['o', 'i'])
 	.alias('i', 'input')
 	.alias('o', 'output')
-	.requiresArg(['o'])
+	.pkgConf('posthtml')
+	.config('c')
+	.alias('c', 'config')
 	.version(function () {
-		return [
-			'posthtml version',
-			require('./node_modules/posthtml/package.json').version
-		].join(' ');
-	}, 'v')
+		return require('./package.json').version;
+	})
 	.alias('v', 'version')
 	.help('h')
 	.alias('h', 'help')
-	.check(function (argv) {
-		if (!argv.input) {
-			throw new Error('Not einter argument --input or -i file');
-		}
-		if (!argv.output) {
-			throw new Error('Not einter argument --output or -o file');
-		}
-		return true;
-	})
 	.argv;
-
 
 // get htmls
 var html = fs.readFileSync(argv.input, 'utf8');